Allow setting up a DNS name for the minecloud server (#70)
* Allow setting up a DNS name for the minecloud server

Implements #68

A DNS name is setup automatically when the parameter
DOMAIN_NAME is set to a name pointing to an existing
Hosted Zone. Setting up a hosted zone on AWS involves
purchasing a domain and paying a for it on a yearly
basis. This makes it hard to automate and is therefore
left out of the automation. Instead an existing
hosted zone is referenced by the value of the parameter
DOMAIN_NAME in the MineCloud-Service-Info.ts file.
When this parameter is set, an additional branch of
the CDK script gets executed, which assigns some
additional policies to the EC2 role to allow the VM
to lookup the hosted zone id to update the IP address
of the ARecord, which has been created by CDK.
The DNS name is derived by appending the value of
DOMAIN_NAME to the hostname "minecloud".
The DOMAIN_NAME is attached as a tag to the EC2 instance
to make it accessible to the start_service.sh script
which updates the ARecord with the assigned IP address.
The bot pushes the domain name to the Discord channel
only when the DOMAIN_NAME parameter has been set.

The changeset also includes a change to avoid the file
MineCloud-Service-Info.ts with all the private data
being committed to Git accidentally. Therefore the
file is renamed to MineCloud-Service-Info.ts.sample
and the README is updated to make a copy to the actual
name MineCloud-Service-Info.ts before setting all the
parameters. The MineCloud-Service-Info.ts is set to the
.gitignore list so that the actual file is ignored
upon git commits.

* Reverts a change that is not ready yet, accidentally committed

* Added some more comments